Update README e algumas correções

+++ src/memalloc.h
@@ -9,6 +9,6 @@ void *dismiss_brk();
 
 void *memory_alloc(unsigned long bytes);
 
-unsigned long int memory_free(void *pointer);
+int memory_free(void *pointer);
 
 #endif
